+++ lib/faker/lorem.rb
@@ -32,7 +32,7 @@ module Faker
         sample('faker.multibyte')
       end
 
-      def sentence(word_count = 4, supplemental = false, random_words_to_add = 6)
+      def sentence(word_count = 4, supplemental = false, random_words_to_add = 0)
         words(word_count + rand(random_words_to_add.to_i), supplemental).join(' ').capitalize + locale_period
       end
 
@@ -40,7 +40,7 @@ module Faker
         1.upto(resolve(sentence_count)).collect { sentence(3, supplemental) }
       end
 
-      def paragraph(sentence_count = 3, supplemental = false, random_sentences_to_add = 3)
+      def paragraph(sentence_count = 3, supplemental = false, random_sentences_to_add = 0)
         sentences(resolve(sentence_count) + rand(random_sentences_to_add.to_i), supplemental).join(locale_space)
       end
 
@@ -56,8 +56,8 @@ module Faker
         paragraph[0...chars - 1] + '.'
       end
 
-      def question(word_count = 4, supplemental = false, random_words_to_add = 6)
-        words(word_count + rand(random_words_to_add.to_i), supplemental).join(locale_space).capitalize + locale_question_mark
+      def question(word_count = 4, supplemental = false, random_words_to_add = 0)
+        words(word_count + rand(random_words_to_add), supplemental).join(' ').capitalize + locale_question_mark
       end
 
       def questions(question_count = 3, supplemental = false)